Support implicit liftToNamespace
Most non-trivial modules will call `liftToNamespace` on
their options, leading to a few lines of boilerplate in each
module.

This change adds support for applying the lift operation by
default to all modules discovered by a given `findModules`
call, which can be overriden on a per-module basis.

While this defaults to disabled, the additional parameter
to `findModules` is still a breaking change.